requirements.txt
groq>=0.4.0
assemblyai>=0.26.0
google-generativeai>=0.7.0
python-dotenv>=1.0.0
